+++ docs/changelog.md
@@ -4,6 +4,14 @@
 
 ## Changelog
 
+### v0.2.0
+
+- Discard log entries with unchanged word count.
+- Center the columns.
+
+Based on PyWriter v12.3.2
+Compatibility: novelyst v4.17 API
+
 ### v0.1.0
 
 - Provide a simple list display.